+++ app/src/js/components/Collections/granules.js
@@ -75,7 +75,7 @@ const CollectionGranules = ({
 
   useEffect(() => {
     setWorkflow(workflowOptions[0]);
-  // eslint-disable-next-line react-hooks/exhaustive-deps
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [JSON.stringify(workflowOptions)]);
 
   function generateQuery() {
@@ -105,7 +105,7 @@ const CollectionGranules = ({
     return bulkActions(granules, actionConfig);
   }
 
-  function selectWorkflow(selector, selectedWorkflow) {
+  function selectWorkflow(_selector, selectedWorkflow) {
     setWorkflow(selectedWorkflow);
   }