+++ app/src/js/components/Granules/list.js
@@ -82,7 +82,7 @@ const AllGranules = ({
 
   useEffect(() => {
     setWorkflow(workflowOptions[0]);
-  // eslint-disable-next-line react-hooks/exhaustive-deps
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [JSON.stringify(workflowOptions)]);
 
   useEffect(() => {